requirements.txt
astroid==3.0.2
contourpy==1.2.0
cycler==0.12.1
dill==0.3.7
fonttools==4.47.2
HTMLTestRunner-rv==1.1.2
isort==5.13.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
mccabe==0.7.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
platformdirs==4.1.0
pylint==3.0.3
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tomlkit==0.12.3
tzdata==2023.4
